+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * use_case selection is flattened alongside device and variant selection
 * Change C build [full] option (now implemented as clean + incremental build)
+* Auto container restart now does not automatically pull the docker image
+
+### Fixed
+
+* Fixed docker container status when VSCode connects to already running image at startup
 
 ## [1.3.2]